♻️ bonus points 분리
