+++ src/basic/features/bonus-points/config.js
@@ -1,0 +1,1 @@
+export const POINT_RATIO = Object.freeze(1000);
